script changed

+++ PA4/scripts/compile.sh
@@ -1,9 +1,15 @@
 #!/bin/bash
 cd src/
-make
-mv cgen ../
 shopt -s extglob
-rm *.o 
-rm !(cgen.cc|cgen.h|Makefile|emit.h|README*|cool-tree.h|cgen_supp.cc)
-cd ..
-clear
+stderr=$(make)
+
+if (echo "$stderr" | grep failed)
+then
+  echo "Compilation failed."
+else
+	mv cgen ../
+	rm *.o 
+	rm !(cgen.cc|cgen.h|Makefile|emit.h|README*|cool-tree.h|cgen_supp.cc)
+	cd ..
+	clear
+fi
